requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.85.0
anyio==4.12.1
attrs==25.4.0
banks==2.4.1
boto3==1.42.69
botocore==1.42.69
cached-property==2.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
chevron==0.14.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.1
filelock==3.25.2
filetype==1.2.0
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.2.0
genanki==0.13.1
google-auth==2.49.1
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
llama-index==0.14.18
llama-index-cli==0.5.6
llama-index-core==0.14.18
llama-index-embeddings-huggingface==0.7.0
llama-index-embeddings-openai==0.6.0
llama-index-indices-managed-llama-cloud==0.11.0
llama-index-instrumentation==0.5.0
llama-index-llms-anthropic==0.11.1
llama-index-llms-openai==0.7.2
llama-index-postprocessor-flag-embedding-reranker==0.5.0
llama-index-readers-llama-parse==0.6.0
llama-index-workflows==2.16.1
llama-parse==0.5.20
llama_cloud==1.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.3
openai==2.29.0
packaging==26.0
pillow==12.1.1
platformdirs==4.9.4
propcache==0.4.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart>=0.0.9
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
wrapt==2.1.2
yarl==1.23.0
